+++ lib/src/object_snapshot.dart
@@ -1,27 +1,31 @@
 part of algolia;
 
+// Implementation & bug solved by https://github.com/franvera
+// - highlightResult [Bug] (commit ref: 0d76d24fe8aa347a0933920afe5ded43bdcbd68b)
+// - snippetResult [Implementation] (commit ref: 0d76d24fe8aa347a0933920afe5ded43bdcbd68b)
+
 class AlgoliaObjectSnapshot {
   Algolia algolia;
   String objectID;
   String index;
-  Map<String, Map<String, dynamic>> highlightResult;
+  Map<String, dynamic> highlightResult;
+  Map<String, dynamic> snippetResult;
   Map<String, dynamic> data;
 
   AlgoliaObjectReference get ref =>
       AlgoliaObjectReference._(algolia, index, objectID);
 
-  // AlgoliaObjectSnapshot.fromMap(Map<String, dynamic> map) {
   AlgoliaObjectSnapshot.fromMap(algolia, index, Map<String, dynamic> map) {
     algolia = algolia;
     index = index;
     objectID = map['objectID'];
-    Map<String, dynamic> newMapHighlightResult = map['_highlightResult'] != null
-        ? Map<String, Map<String, dynamic>>.from(map['_highlightResult'])
-        : null;
-    highlightResult = newMapHighlightResult;
-    // Map<String, dynamic> m = map;
+
+    highlightResult = map['_highlightResult'];
+    snippetResult = map['_snippetResult'];
+
     map.remove('objectID');
     map.remove('_highlightResult');
+    map.remove('_snippetResult');
     data = map;
   }
 }